requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Validation
email-validator==2.1.0.post1

# HTTP Client (for external API calls)
httpx==0.26.0

# CORS
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# OCR and Image Processing
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# SQLite async support
aiosqlite==0.19.0
